Modifications to prevent us from harrassing ourselves.
Not as schizophrenic as it sounds.  These changes are to save us from
having the bioconductor@stat.math.ethz.ch email added to the list of
people who are not subscribed to bioc-devel.  This just checks the
latest list of emails, and then adds this email to that list (as if it
were subscribed) so that I don't have this email coming up on the
outputs of people who are not subscribed.