Update adds sqlShemaGenerators file.
This update adds the schemaGenerators file which will hold functions
to make packges based on all the different kinds of schemas we will
support.  There are also notes added to the test.R script that detail
my next move which will have to be to fix a bug that makes things a
little too beholden to the chip based packages (when tallying
map_counts).  The helper functions have to be a teeny bit better about
what they construe as the primary table for a schema.