Finished cleaning up for MOUSE_DB
Changes correct the map_counts and other metadata.  These changes
include a small refactor that will make the code easier to maintain
and allow the functions to be more flexible in the future as we add
other kinds of schemas.  Basically these move the substitution
parameters out into a list passed in from the parent function.