Script outputs summary table.
This adds a primitive summary matrix as an output so that I can look
at the miRNA seeds (now not redundant) as rows and the UTRs as cols,
(with indications for what ones were from sources or the seed group).
And then it shows how many times for each element that particular seed
was found.  Seeds that were not found are not listed.