This fixes a bad bug in how the individual elements are sorted.
So this algorithm basically permutes through all possible arrangements
for each of a paired set of scores and then needs to assess which
strings of numbers look the most alike.  Before I was using the
euclidean distance as a stopgap, but this should be more sensible.
NAs are still ignored, but at least now the order that they happen in
should matter more, which is key to getting things sorted correctly.