Major bug fixes.
These bug fixes are for speeding up the rate at which the script can
work through the permutations and another fix which corrects a bug
that is uncovered by not having equal sized permuation matrices
anymore.  The speed up is achieved by prefiltering out any
permutations that are not ordered (which can be most of them for lists
where there is a lot of numerical values).  This uncovered another bug
where I was improperly recycling an index.