This fix adds support for using gaps (extra NAs) in the assembly.
So when we are figuring out how to align the utative seed regions, we
can use gaps, but using more gaps takes a LOT more permutations to
sort out, so it can slow things way down.  Most of these are SUPER
short, and so gaps are not needed, so the default is 1.