New version of script does not bother biomaRt.
Mostly this version does not bother ME, because it does not ask
biomaRt all the time whether or not it has a UTR, instead, it just
uses some data files that I made from all the UTRs that I could get
from biomaRt, (just went down the list of EGs and pulled everything
down).  So now I don't get nearly as many errors when I run against a
random sample.