Added some code to replace some HORRIBLY repetitive SQL.
Also, now ALL of the inparanoid tables are present in
inparanoid.sqlite.  So this kills two birds with one stone.  BUT, the
downstream SQL will have to be checked carefully since it might be
expecting that some tables will be named in the opposite direction
(ie. homsa_musmu instead of musmu_homsa for example).