Adds yeastNCBI database.
This adds a yeast database that is sourced from NCBI.  The reasons for
this are so that we can finally get the people who are upset about the
lack of an entrez gene ID for yeast to SHUT UP.  Also, Tony needs both
entrez gene IDs, and UNIPROT IDs and this is the only way (so far)
that I can find to get all these IDs in one place (along with standard
locus IDs and SGD IDs which this database ALSO includes.  Some care
also had to be taken to get KEGG and GO to map correctly into this
package.  KEGG in particular would not map correctly until I had the
locus IDs in place, and I needed those anyhow so these things all came
together.